refactor: 자동차 경주 검증 로직을 racingGameValidator에서 Cars로 이동
