+++ src/main/java/racingcar/domain/Cars.java
@@ -3,18 +3,21 @@ package racingcar.domain;
 import static racingcar.Util.Utils.*;
 
 import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.HashSet;
 import java.util.List;
+import java.util.Set;
 import java.util.stream.Collectors;
 
-import racingcar.validator.RacingGameValidator;
-
 public class Cars {
+	public static final String ERROR_ONLY_NAME = "[ERROR] 이름을 2개 이상 입력해주세요.";
+	public static final String ERROR_DUPLICATE_NAME = "[ERROR] 중복된 이름이 있습니다.";
 	public static final int MOVE_CONDITION = 4;
 
 	private final List<Car> cars = new ArrayList<>();
 
 	public Cars(String[] carNames) {
-		RacingGameValidator.validateCarNames(carNames);
+		validateCarNames(carNames);
 		for (String carName : carNames) {
 			cars.add(new Car(carName));
 		}
@@ -22,6 +25,24 @@ public class Cars {
 
 	public Cars(List<Car> cars) {
 		this.cars.addAll(cars);
+	}
+
+	private void validateCarNames(String[] names) {
+		checkDuplicateName(names);
+		checkOnlyName(names);
+	}
+
+	private static void checkOnlyName(String[] names) {
+		if (names.length == 1) {
+			throw new IllegalArgumentException(ERROR_ONLY_NAME);
+		}
+	}
+
+	private static void checkDuplicateName(String[] names) {
+		Set<String> hashNames = new HashSet<>(Arrays.asList(names));
+		if (hashNames.size() != names.length) {
+			throw new IllegalArgumentException(ERROR_DUPLICATE_NAME);
+		}
 	}
 
 	public List<Car> getCars() {
