+++ src/test/java/racingcar/domain/CarsTest.java
@@ -1,13 +1,16 @@
 package racingcar.domain;
 
+import static org.assertj.core.api.Assertions.*;
+import static racingcar.domain.Cars.*;
+
+import java.util.Arrays;
+import java.util.List;
+
+import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import static org.assertj.core.api.Assertions.assertThat;
-
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
+import racingcar.Util.Utils;
 
 class CarsTest {
 	@Test
@@ -16,6 +19,22 @@ class CarsTest {
 		String[] carNames = {"pobi", "crong", "huan"};
 		Cars cars = new Cars(carNames);
 		assertThat(cars.getCars().size()).isEqualTo(carNames.length);
+	}
+
+	@Test
+	@DisplayName("입력한 이름값이 한개인 경우")
+	void onlyName() {
+		Assertions.assertThatThrownBy(() -> new Cars(Utils.splitter("pobi")))
+			.isInstanceOf(IllegalArgumentException.class)
+			.hasMessageContaining(ERROR_ONLY_NAME);
+	}
+
+	@Test
+	@DisplayName("중복된 이름이 있을 경우")
+	void duplicateName() {
+		Assertions.assertThatThrownBy(() -> new Cars(Utils.splitter("pobi,pobi")))
+			.isInstanceOf(IllegalArgumentException.class)
+			.hasMessageContaining(ERROR_DUPLICATE_NAME);
 	}
 
 	@Test
